requirements.txt
python-docx
openai
llama-index
python-dotenv